Refactor routing and component structure; rename ShopDetail to RestaurantDetails and update imports
