+++ src/App.jsx
@@ -5,17 +5,12 @@ import Header from '@src/components/Header';
 import Footer from "@src/components/Footer";
 import Home from "./pages/Home";
 
-import Shop from "./pages/Shop/Shop";
-import ShopDetail from "./pages/Shop/ShopDetail";
+import RestaurantDetails from "./pages/RestaurantDetails.jsx";
 
-import ShopCart from "./pages/Shop/ShopCart";
-import ShopCheckOut from "./pages/Shop/ShopCheckOut";
-import MyAccountOrder from "./pages/Accounts/MyAccountOrder";
-import MyAccountSetting from "./pages/Accounts/MyAcconutSetting";
+
 import AddressSelection from "./pages/AddressSelection.jsx";
-import MyAccountForgetPassword from "./pages/Accounts/MyAccountForgetPassword";
-import Login from "./pages/Accounts/login.jsx";
-import Register from "./pages/Accounts/register.jsx";
+import Login from "./pages/login.jsx";
+import Register from "./pages/register.jsx";
 import ErrorPage from "@src/pages/ErrorPage";
 import GlobalResetContext from "@src/context/GlobalResetContext";
 
@@ -29,14 +24,8 @@ function App() {
 
                     <Routes key={resetKey}>
                         <Route path="/" element={<Home />} />
-                        <Route path="/Restaurants" element={<Shop />} />
-                        <Route path="/Restaurant/:id" element={<ShopDetail />} />
-                        <Route path="/CheckOut" element={<><ShopCheckOut /></>} />
-                        <Route path="/Cart" element={<><ShopCart /></>} />
-                        <Route path="/MyAccountOrder" element={<><MyAccountOrder /></>} />
-                        <Route path="/MyAccountSetting" element={<><MyAccountSetting /></>} />
+                        <Route path="/Restaurant/:id" element={<RestaurantDetails />} />
                         <Route path="/Address" element={<><AddressSelection /></>} />
-                        <Route path="/MyAccountForgetPassword" element={<MyAccountForgetPassword />} />
                         <Route path="/login" element={<Login />} />
                         <Route path="/Register" element={<Register />} />
